reformat null object.
